@@ -7,15 +7,15 @@ import android.graphics.Paint;
 import android.graphics.PorterDuff.Mode;
 import android.graphics.PorterDuffXfermode;
 import android.graphics.RectF;
-import android.support.v7.widget.AppCompatImageView;
+import androidx.appcompat.widget.AppCompatImageView;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
 
 /**
-* creat_user: zhengzaihong
+* create_user: zhengzaihong
 * Email:1096877329@qq.com
-* creat_date: 2017/12/6
-* creat_time: 10:28
+* create_date: 2017/12/6
+* create_time: 10:28
 * describe:支持圆角的view
 **/
 public class RoundIconView extends AppCompatImageView {
@@ -36,7 +36,7 @@ public class RoundIconView extends AppCompatImageView {
 
 
     private final RectF roundRect = new RectF();
-    private float rect_adius = 0;
+    private float rect_radius = 0;
     private final Paint maskPaint = new Paint();
     private final Paint zonePaint = new Paint();
  
@@ -46,11 +46,11 @@ public class RoundIconView extends AppCompatImageView {
         zonePaint.setAntiAlias(true);
         zonePaint.setColor(Color.WHITE);
         float density = getResources().getDisplayMetrics().density;
-        rect_adius = rect_adius * density;
+        rect_radius = rect_radius * density;
     }
  
-    public void setRectAdius(float adius) {
-        rect_adius = adius;
+    public void setRectRadius(float adius) {
+        rect_radius = adius;
         invalidate();
     }
  
@@ -66,7 +66,7 @@ public class RoundIconView extends AppCompatImageView {
     @Override
     public void draw(Canvas canvas) {
         canvas.saveLayer(roundRect, zonePaint, Canvas.ALL_SAVE_FLAG);
-        canvas.drawRoundRect(roundRect, rect_adius, rect_adius, zonePaint);
+        canvas.drawRoundRect(roundRect, rect_radius, rect_radius, zonePaint);
         //
         canvas.saveLayer(roundRect, maskPaint, Canvas.ALL_SAVE_FLAG);
         super.draw(canvas);
