@@ -1,15 +1,24 @@
 package com.dz.ninegridimage;
 
+import android.app.ActivityManager;
 import android.app.Application;
+import android.content.ComponentName;
+import android.content.Context;
 import android.graphics.Color;
+import android.os.Build;
+import android.util.Log;
 
-import com.dz.utlis.JavaUtils;
 import com.dz.utlis.ScreenUtils;
 import com.dz.utlis.ToastTool;
 import com.dz.utlis.UiCompat;
-import com.dz.utlis.view.ToastConfig;
 
-import static com.dz.utlis.JavaUtils.*;
+import java.lang.reflect.Method;
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+
+import static com.dz.utlis.JavaUtils.isdebug;
 
 
 public class App extends Application {
@@ -20,19 +29,19 @@ public class App extends Application {
 
         isdebug = true;
         // Toast 配置
-        ToastConfig config = new ToastConfig()
-            .setInterval(2000)
-            .setRadiusBg((int) ScreenUtils.dip2px(this,30))
-            .setToastTextColor(UiCompat.getColor(getResources(),R.color.light_blue_200))
-            .setToastViewGroupBgColor(Color.WHITE)
-            .setToastTextSize(16)
-            .setBgPadding((int) ScreenUtils.dip2px(this,15))
-            .setShortToast(false)
-            .setStrokeWidth(0)
-            .setRadiusType(ToastConfig.RadiusType.ALL_RADIUS)
-            .setStrokeColor(Color.GREEN);
+        ToastTool.options()
+                .setInterval(2000)
+                .setRadius((int) ScreenUtils.dip2px(this, 30))
+                .setTextColor(UiCompat.getColor(getResources(), R.color.light_blue_200))
+                .setBackGroundColor(Color.WHITE)
+                .setTextSize(16)
+                .setPadding((int) ScreenUtils.dip2px(this, 15))
+                .setLongTime(false)
+                .setStrokeWidth(0)
+                .setRadiusType(ToastTool.RadiusType.ALL_RADIUS)
+                .setStrokeColor(Color.GREEN)
+                .build(this);
 
-        //初始化 Toast工具
-        ToastTool.get().initConfig(this,config);
+
     }
 }
