@@ -3,6 +3,8 @@ package com.dz.ninegridimages.interfaces;
 import android.content.Context;
 import android.widget.ImageView;
 
+import java.io.Serializable;
+
 /**
  * creat_user: zhengzaihong
  * email:1096877329@qq.com
@@ -11,17 +13,42 @@ import android.widget.ImageView;
  * describe: 图片加载接口
  **/
 
-public interface ImageLoader {
+public interface ImageLoader extends Serializable {
+
+
     /**
-     * 需要子类实现该方法，以确定如何加载和显示图片
-     *
-     * @param context   上下文
-     * @param imageView 需要展示图片的ImageView
-     * @param object    实体类
+     * 显示宫格视图的 监听器
      */
-    <T> void displayImage(Context context, ImageView imageView, T object);
+    interface OnNineGridImageListener extends ImageLoader{
+
+        /**
+         * 需要子类实现该方法，以确定如何加载和显示图片
+         *
+         * @param context   上下文
+         * @param imageView 需要展示图片的ImageView
+         * @param obj    实体类
+         */
+
+        <T> void displayImage(Context context, ImageView imageView, T obj);
+
+    }
 
 
-    <T> void loadPreImage(Context context, ImageView imageView, T object);
+    /**
+     * 加载预览大图的监听器
+     */
+    interface OnPreBigImageListener extends ImageLoader{
+
+        /**
+         * 需要子类实现该方法，以确定如何加载和显示图片
+         *
+         * @param context   上下文
+         * @param imageView 需要展示图片的ImageView
+         * @param obj    实体类
+         */
+        <E> void loadPreImage(Context context, ImageView imageView, E obj,int index);
+
+    }
+
 
 }
