@@ -1,9 +1,10 @@
 package com.dz.ninegridimages.preview;
 
 import android.content.Context;
-import android.support.v4.view.ViewPager;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
+
+import androidx.viewpager.widget.ViewPager;
 
 /** 修复图片在ViewPager控件中缩放报错的BUG */
 public class HackyViewPager extends ViewPager {
@@ -25,6 +26,7 @@ public class HackyViewPager extends ViewPager {
         }
         return false;
     }
+
 
     @Override
     public boolean onInterceptTouchEvent(MotionEvent ev) {
